requirements.txt
langgraph>=0.2.0
langchain-anthropic>=0.3.0
langchain-core>=0.3.0
anthropic>=0.40.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.2.0
pydantic>=2.7.0
python-dotenv>=1.0.0
tavily-python>=0.3.0
google-search-results>=2.4.2
fastapi>=0.115.0
uvicorn>=0.30.0
playwright==1.44.0
aiofiles>=23.2.0
pyyaml>=6.0.1
pytest>=8.0.0
